sonar fixes and refactor
