+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/IFunction.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace AWS.Lambda.Powertools.JMESPath.Functions;
+
+internal interface IFunction
+{
+    int? Arity { get; }
+    bool TryEvaluate(DynamicResources resources, IList<IValue> args, out IValue element);
+}
